resolve comments

+++ internal/database/schema.go
@@ -59,8 +59,8 @@ type Report struct {
 	IsUpload       bool
 
 	CreationTime       time.Time
-	CompletedFileCount uint64 `gorm:"default:0"`
-	TotalFileCount     uint64 `gorm:"default:0"`
+	CompletedFileCount int `gorm:"default:0"`
+	TotalFileCount     int `gorm:"default:0"`
 
 	Tags       []ReportTag `gorm:"foreignKey:ReportId;constraint:OnDelete:CASCADE"`
 	CustomTags []CustomTag `gorm:"foreignKey:ReportId;constraint:OnDelete:CASCADE"`
